requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
result==0.16.1
returns==0.22.0
sqlparse==0.4.4
typing==3.7.4.3
typing_extensions==4.10.0
